# api/admin/v1/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "unauthenticated",
              "forbidden",
              "unavailable",
              "internal",
              "not_found",
              "method_not_allowed",
              "invalid_content_type",
              "invalid_body",
              "invalid_request",
              "invalid_idempotency_key",
              "precondition_required",
              "invalid_precondition",
              "precondition_failed",
              "idempotency_conflict",
              "invalid_reference",
              "unsupported",
              "validation_failed",
              "image_unavailable",
              "api_unavailable",
              "commit_unknown"
            ],
            "type": "string"
          },
          "operationID": {
            "maxLength": 63,
            "minLength": 1,
            "pattern": "^[a-z0-9](?:[-a-z0-9]{0,61}[a-z0-9])?$",
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "suggestedAction": {
            "maxLength": 512,
            "type": "string"
          },
          "version": {
            "enum": [
              "v1"
            ],
            "maxLength": 512,
            "type": "string"
          }
        },
        "required": [
          "version",
          "code"
        ],
        "type": "object"
      },
...
      "Self": {
        "additionalProperties": false,
        "properties": {
          "credentialId": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.-]+$",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "maxLength": 512,
            "type": "string"
          },
          "namespace": {
            "maxLength": 63,
            "minLength": 1,
            "pattern": "^[a-z0-9](?:[-a-z0-9]{0,61}[a-z0-9])?$",
            "type": "string"
          },
          "principalId": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.-]+$",
            "type": "string"
          },
          "version": {
            "enum": [
              "v1"
            ],
            "maxLength": 512,
            "type": "string"
          }
        },
        "required": [
          "version",
          "principalId",
          "credentialId",
          "expiresAt",
          "namespace"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Projected opaque administrator credential; authenticated before all validation, reads and mutation effects.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Arcadectl ordinary administrator API",
    "version": "v1"
  },
  "openapi": "3.0.3",
  "paths": {
    "/v1/auth/self": {
      "get": {
        "operationId": "auth_self__v1_auth_self",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Self"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Versioned bounded response; no raw worker, credential or Kubernetes error output."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-arcadectl-action": "auth.self"
      }
    },
...
  }
}